+++ resources/views/partials/article.blade.php
@@ -1,9 +1,8 @@
 @php( $search_tags = display_search_keywords( get_the_ID() ) )
 
 @if (get_post_format( $p->ID ) === 'image')
-    <div class="pt-1 cf-article" data-label="{{ $search_tags }}" style="{{ display_styles(get_the_ID()) }} display:inline-block; overflow:scroll">
+    <div data-title="{{ the_title() }}" class="pt-1 cf-article cf-image" data-label="{{ $search_tags }}" style="{{ display_styles(get_the_ID()) }} display:inline-block; overflow:scroll">
         <a data-toggle="modal" data-target=".image-zoom">
-            <h3>{{ the_title()}}</h3>
             @php(the_content())
         </a>
         <!-- Modal -->
@@ -17,7 +16,19 @@
     </div>
 @else
     <div class="p-1 cf-article" data-label="{{ $search_tags }}" style="{{ display_styles(get_the_ID()) }} display:inline-block; overflow:scroll">
-        <h3>{{ the_title()}}</h3>
         @php( the_content() )
     </div>
 @endif
+
+
+<script type="text/javascript">
+jQuery(document).ready(function() {
+    $('.cf-image').hover(
+    function(){
+        var title = $(this).attr("data-title")
+        $(".footer-panel").html(title).slideDown(300);
+    }, function(){
+        $(".footer-panel").slideUp(300);
+    });
+});
+</script>